xorg: hidp screen and ubuntu gdm support for i3
After migrating to Ubuntu 20.04 because of my new job, I've had to do
some minor tweaks to the xorg config in order to make i3 work as it used
to.

Basically, .xinitrc is no longer used because gdm (no startx anymore).
This leads me to put .Xresources in its expected path so it's
automatically loaded.

Also, .xsessionrc is Debian/Ubuntu specific and allows for further
customzation (wm-independent).

Finally, because I'm using a hidp screen now, I've increased the dpi to
150% and the mouse cursor.

References:
- http://unix.stackexchange.com/questions/281858/difference-between-xinitrc-xsession-and-xsessionrc
- https://wiki.debian.org/Xsession
- https://www.reddit.com/r/i3wm/comments/8ix88z/how_can_i_change_the_mouse_cursor_size/